+++ src/models/user_info_base.ts
@@ -1,4 +1,0 @@
-
-export interface UserInfoBase {
-    _type: 'personal' | 'enterprise' | 'iosEnterprise'
-}